refactor(pagerduty): move configuration types into pagerduty package
