+++ notify/pagerduty/pagerduty_test.go
@@ -33,7 +33,6 @@ import (
 
 	amcommoncfg "github.com/prometheus/alertmanager/config/common"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/notify/test"
 	"github.com/prometheus/alertmanager/template"
@@ -42,7 +41,7 @@ import (
 
 func TestPagerDutyRetryV1(t *testing.T) {
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			ServiceKey: commoncfg.Secret("01234567890123456789012345678901"),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -60,7 +59,7 @@ func TestPagerDutyRetryV1(t *testing.T) {
 
 func TestPagerDutyRetryV2(t *testing.T) {
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -82,7 +81,7 @@ func TestPagerDutyRedactedURLV1(t *testing.T) {
 
 	key := "01234567890123456789012345678901"
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			ServiceKey: commoncfg.Secret(key),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -101,7 +100,7 @@ func TestPagerDutyRedactedURLV2(t *testing.T) {
 
 	key := "01234567890123456789012345678901"
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			URL:        &amcommoncfg.URL{URL: u},
 			RoutingKey: commoncfg.Secret(key),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
@@ -125,7 +124,7 @@ func TestPagerDutyV1ServiceKeyFromFile(t *testing.T) {
 	defer fn()
 
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			ServiceKeyFile: f.Name(),
 			HTTPConfig:     &commoncfg.HTTPClientConfig{},
 		},
@@ -149,7 +148,7 @@ func TestPagerDutyV2RoutingKeyFromFile(t *testing.T) {
 	defer fn()
 
 	notifier, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			URL:            &amcommoncfg.URL{URL: u},
 			RoutingKeyFile: f.Name(),
 			HTTPConfig:     &commoncfg.HTTPClientConfig{},
@@ -176,23 +175,23 @@ func TestPagerDutyTemplating(t *testing.T) {
 
 	for _, tc := range []struct {
 		title string
-		cfg   *config.PagerdutyConfig
+		cfg   *PagerdutyConfig
 
 		retry  bool
 		errMsg string
 	}{
 		{
 			title: "full-blown legacy message",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
-				Images: []config.PagerdutyImage{
+				Images: []PagerdutyImage{
 					{
 						Src:  "{{ .Status }}",
 						Alt:  "{{ .Status }}",
 						Href: "{{ .Status }}",
 					},
 				},
-				Links: []config.PagerdutyLink{
+				Links: []PagerdutyLink{
 					{
 						Href: "{{ .Status }}",
 						Text: "{{ .Status }}",
@@ -208,16 +207,16 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "full-blown legacy message",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
-				Images: []config.PagerdutyImage{
+				Images: []PagerdutyImage{
 					{
 						Src:  "{{ .Status }}",
 						Alt:  "{{ .Status }}",
 						Href: "{{ .Status }}",
 					},
 				},
-				Links: []config.PagerdutyLink{
+				Links: []PagerdutyLink{
 					{
 						Href: "{{ .Status }}",
 						Text: "{{ .Status }}",
@@ -233,7 +232,7 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "nested details",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 				Details: map[string]any{
 					"a": map[string]any{
@@ -251,7 +250,7 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "nested details with template error",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 				Details: map[string]any{
 					"a": map[string]any{
@@ -267,7 +266,7 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "details with templating errors",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 				Details: map[string]any{
 					"firing":       `{{ .Alerts.Firing | toJson`,
@@ -280,7 +279,7 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "v2 message with templating errors",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 				Severity:   "{{ ",
 			},
@@ -288,7 +287,7 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "v1 message with templating errors",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				ServiceKey: commoncfg.Secret("01234567890123456789012345678901"),
 				Client:     "{{ ",
 			},
@@ -296,14 +295,14 @@ func TestPagerDutyTemplating(t *testing.T) {
 		},
 		{
 			title: "routing key cannot be empty",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				RoutingKey: commoncfg.Secret(`{{ "" }}`),
 			},
 			errMsg: "routing key cannot be empty",
 		},
 		{
 			title: "service_key cannot be empty",
-			cfg: &config.PagerdutyConfig{
+			cfg: &PagerdutyConfig{
 				ServiceKey: commoncfg.Secret(`{{ "" }}`),
 			},
 			errMsg: "service key cannot be empty",
@@ -398,7 +397,7 @@ func TestEventSizeEnforcement(t *testing.T) {
 	}
 
 	notifierV1, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			ServiceKey: commoncfg.Secret("01234567890123456789012345678901"),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -421,7 +420,7 @@ func TestEventSizeEnforcement(t *testing.T) {
 	}
 
 	notifierV2, err := New(
-		&config.PagerdutyConfig{
+		&PagerdutyConfig{
 			RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -445,7 +444,7 @@ func TestPagerDutyEmptySrcHref(t *testing.T) {
 		Links       []pagerDutyLink
 	}
 
-	images := []config.PagerdutyImage{
+	images := []PagerdutyImage{
 		{
 			Src:  "",
 			Alt:  "Empty src",
@@ -463,7 +462,7 @@ func TestPagerDutyEmptySrcHref(t *testing.T) {
 		},
 	}
 
-	links := []config.PagerdutyLink{
+	links := []PagerdutyLink{
 		{
 			Href: "",
 			Text: "Empty href",
@@ -479,11 +478,7 @@ func TestPagerDutyEmptySrcHref(t *testing.T) {
 		if image.Src == "" {
 			continue
 		}
-		expectedImages = append(expectedImages, pagerDutyImage{
-			Src:  image.Src,
-			Alt:  image.Alt,
-			Href: image.Href,
-		})
+		expectedImages = append(expectedImages, pagerDutyImage(image))
 	}
 
 	expectedLinks := make([]pagerDutyLink, 0, len(links))
@@ -533,7 +528,7 @@ func TestPagerDutyEmptySrcHref(t *testing.T) {
 	url, err := url.Parse(server.URL)
 	require.NoError(t, err)
 
-	pagerDutyConfig := config.PagerdutyConfig{
+	pagerDutyConfig := PagerdutyConfig{
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 		RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 		URL:        &amcommoncfg.URL{URL: url},
@@ -601,7 +596,7 @@ func TestPagerDutyTimeout(t *testing.T) {
 			u, err := url.Parse(srv.URL)
 			require.NoError(t, err)
 
-			cfg := config.PagerdutyConfig{
+			cfg := PagerdutyConfig{
 				HTTPConfig: &commoncfg.HTTPClientConfig{},
 				RoutingKey: commoncfg.Secret("01234567890123456789012345678901"),
 				URL:        &amcommoncfg.URL{URL: u},
@@ -872,7 +867,7 @@ func TestRenderDetails(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			n := &Notifier{
-				conf: &config.PagerdutyConfig{
+				conf: &PagerdutyConfig{
 					Details: tt.args.details,
 				},
 				tmpl: test.CreateTmpl(t),
